requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
clip-openai==1.0.post20230121
datasets==2.10.1
dill==0.3.6
filelock==3.9.0
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
huggingface-hub==0.13.1
idna==3.4
importlib-metadata==6.0.0
multidict==6.0.4
multiprocess==0.70.14
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
open-clip-torch==2.16.0
packaging==23.0
pandas==1.3.5
Pillow==9.4.0
protobuf==3.20.3
pyarrow==11.0.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
responses==0.18.0
sentencepiece==0.1.97
six==1.16.0
timm==0.6.12
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
xxhash==3.2.0
yarl==1.8.2
zipp==3.15.0
